Changed CalcMortgagePayment types to float64

+++ lib.go
@@ -2,11 +2,11 @@ package finlib
 
 import "math"
 
-func CalcMortgagePayment(loanAmt int, numOfMonths int, interestRate float64) float64 {
+func CalcMortgagePayment(loanAmt float64, numOfMonths float64, interestRate float64) float64 {
 
 	interestRatePerMonth := (interestRate / 100) / 12
 
-	q := math.Pow(1+interestRatePerMonth, float64(numOfMonths))
+	q := math.Pow(1+interestRatePerMonth, numOfMonths)
 
-	return float64(loanAmt) * (interestRatePerMonth * q) / (q - 1)
+	return loanAmt * interestRatePerMonth * q / (q - 1)
 }
